+++ include/ignition/transport/NodeOptions.hh
@@ -126,9 +126,18 @@ namespace ignition
       public: bool TopicRemap(const std::string &_fromTopic,
                               std::string &_toTopic) const;
 
+#ifdef _WIN32
+// Disable warning C4251 which is triggered by
+// std::unique_ptr
+#pragma warning(push)
+#pragma warning(disable: 4251)
+#endif
       /// \internal
       /// \brief Smart pointer to private data.
       protected: std::unique_ptr<transport::NodeOptionsPrivate> dataPtr;
+#ifdef _WIN32
+#pragma warning(pop)
+#endif
     };
     }
   }