+++ log/include/ignition/transport/log/Descriptor.hh
@@ -94,13 +94,21 @@ namespace ignition
         /// \internal Implementation for this class
         class Implementation;
 
+#ifdef _WIN32
+// Disable warning C4251 which is triggered by
+// std::*
+#pragma warning(push)
+#pragma warning(disable: 4251)
+#endif
         /// \internal Pointer to implementation
         private: std::unique_ptr<Implementation> dataPtr;
+#ifdef _WIN32
+#pragma warning(pop)
+#endif
       };
       }
     }
   }
 }
-
 
 #endif