+++ log/include/ignition/transport/log/Log.hh
@@ -104,11 +104,32 @@ namespace ignition
         public: Batch QueryMessages(
             const QueryOptions &_options = AllTopics());
 
+        /// \brief Get start time of the log, or in other words the
+        /// time of the first message found in the log
+        /// \return start time of the log, or zero if the log is not
+        /// valid or if data retrieval failed.
+        public: std::chrono::nanoseconds StartTime() const;
+
+        /// \brief Get end time of the log, or in other words the
+        /// time of the last message found in the log
+        /// \return end time of the log, or zero if the log is not
+        /// valid or if data retrieval failed.
+        public: std::chrono::nanoseconds EndTime() const;
+
         /// \internal Implementation for this class
         private: class Implementation;
 
+#ifdef _WIN32
+// Disable warning C4251 which is triggered by
+// std::*
+#pragma warning(push)
+#pragma warning(disable: 4251)
+#endif
         /// \brief Private implementation
         private: std::unique_ptr<Implementation> dataPtr;
+#ifdef _WIN32
+#pragma warning(pop)
+#endif
       };
       }
     }