+++ test/integration/scopedTopicSubscriber_aux.cc
@@ -26,7 +26,7 @@
 using namespace ignition;
 
 static bool cbExecuted;
-static std::string g_topic = "/foo";
+static std::string g_topic = "/foo"; // NOLINT(*)
 static int data = 5;
 
 //////////////////////////////////////////////////