+++ test/integration/twoProcessesPubSub.cc
@@ -26,10 +26,10 @@
 
 using namespace ignition;
 
-static std::string partition;
-static std::string g_FQNPartition;
-static const std::string g_topic = "/foo";
-static std::string data = "bar";
+static std::string partition;  // NOLINT(*)
+static std::string g_FQNPartition;  // NOLINT(*)
+static const std::string g_topic = "/foo";  // NOLINT(*)
+static std::string data = "bar";  // NOLINT(*)
 static bool cbExecuted = false;
 static bool cbInfoExecuted = false;
 static bool genericCbExecuted = false;
@@ -86,7 +86,7 @@ void cbVector(const ignition::msgs::Vector3d &/*_msg*/)
 }
 
 //////////////////////////////////////////////////
-void cbRaw(const char * /*_msgData*/, const int /*_size*/,
+void cbRaw(const char * /*_msgData*/, const size_t /*_size*/,
            const ignition::transport::MessageInfo &/*_info*/)
 {
   cbRawExecuted = true;
@@ -295,19 +295,19 @@ TEST(twoProcPubSub, PubSubWrongTypesTwoRawSubscribers)
   bool correctRawCbExecuted = false;
   bool genericRawCbExecuted = false;
 
-  auto wrongCb = [&](const char *, const int /*_size*/,
+  auto wrongCb = [&](const char *, const size_t /*_size*/,
                      const transport::MessageInfo &)
   {
     wrongRawCbExecuted = true;
   };
 
-  auto correctCb = [&](const char *, const int /*_size*/,
+  auto correctCb = [&](const char *, const size_t /*_size*/,
                        const transport::MessageInfo &)
   {
     correctRawCbExecuted = true;
   };
 
-  auto genericCb = [&](const char *, const int /*_size*/,
+  auto genericCb = [&](const char *, const size_t /*_size*/,
                        const transport::MessageInfo &)
   {
     genericRawCbExecuted = true;